greybus: spi: working
This change makes a number of cleanup modifications to transport-tcpip.c
but the main change is in greybus/spi.c in gb_spi_protocol_transfer().

The main issue was due to the buffer strategy being quite different
between NuttX (where the original code came from) and Zephyr.

In Zephyr, for general SPI transfers, tx_buf and rx_buf need to be
the same length. The memory locations that they point to and the length
of those memory locations can be NULL / 0.

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>