subsys: greybus: allow multiple drivers on same bundle
allow multiple cport protocol driver of the same bundle
class on the same bundle, this allows GPIO, I2C, SPI
to be on the same gbphy bundle and will make fetching
resources at socket(Eg. mikroBUS) level easier at the
host kernel.

TODO: this just works,but there seems to be an overwrite
bundle->dev on each driver init, but somehow it does not
affect functionality.

Signed-off-by: Vaishnav M A <vaishnav@beagleboard.org>