include: dt-bindings: pinctrl cc13xx_cc26xx fix whitespace errors
this header is a trimmed down version of
hal/ti/.../cc13x2_cc26x2/driverlib/ioc.h
which helps in writing the pinctrl DT nodes in a reusable manner.

The initial commit adds the necessary sections from hal/driverlib
and this commit only fixes the whitespace/styling issues with
the original source.

Signed-off-by: Vaishnav Achath <vaishnav@beagleboard.org>